+++ src/main/java/io/github/tufkan1/projectex/ProjectEX.java
@@ -4,6 +4,8 @@ import io.github.tufkan1.projectex.api.emc.EmcApi;
 import io.github.tufkan1.projectex.command.EmcCommands;
 import io.github.tufkan1.projectex.content.ProjectEXMenus;
 import io.github.tufkan1.projectex.content.ProjectEXBlocks;
+import io.github.tufkan1.projectex.content.ProjectEXComponents;
+import io.github.tufkan1.projectex.content.ProjectEXItems;
 import io.github.tufkan1.projectex.emc.mapping.minecraft.MinecraftRecipeMappingService;
 import io.github.tufkan1.projectex.emc.reload.EmcDataReloadListener;
 import io.github.tufkan1.projectex.internal.emc.EmcValueRegistry;
@@ -28,7 +30,9 @@ public final class ProjectEX implements ModInitializer {
             id("emc_values"),
             new EmcDataReloadListener(EMC_VALUES)
         );
+        ProjectEXComponents.register();
         ProjectEXBlocks.register();
+        ProjectEXItems.register();
         ProjectEXMenus.register();
         EmcCommands.register();
         AlchemyNetworking.register();